requirements.txt
langchain>=0.3
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
langchain-text-splitters
gradio
pypdf
sentence-transformers
pymupdf
python-dotenv
